**chore: remove unused Neo4j repository methods and tests**
- Deleted `CreateDebtItem`, `UpdateStatus`, `CreateService`, `UpdateService`, and `UpdateTeam` methods along with their associated test files.
- Cleaned up related imports and references across the codebase.
